requirements.txt
# App & server
Flask>=3.0,<4
flask-cors>=4.0,<5
gunicorn>=23.0,<24

# LINE SDK
line-bot-sdk>=3.11,<4

# Google Cloud
firebase-admin>=6.6,<7
google-cloud-firestore>=2.16,<3
google-cloud-secret-manager>=2.20,<3
google-cloud-storage>=3.3,<4
google-cloud-pubsub>=2.20,<3

# Utils
python-dateutil>=2.9.0.post0,<3
requests>=2.31,<3
reportlab>=4.2,<5

# JWT for magic link / LINE Login (RS256)
PyJWT[crypto]>=2.9.0,<3
matplotlib==3.9.*
WeasyPrint>=62.0,<63
Jinja2>=3.1,<4